refactor: CouponDto에서 CouponFactory로 이름 수정
