+++ backend/src/test/java/com/woowacourse/kkogkkog/member/domain/repository/MemberHistoryRepositoryTest.java
@@ -1,6 +1,6 @@
 package com.woowacourse.kkogkkog.member.domain.repository;
 
-import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFixture.createCoupon;
+import static com.woowacourse.kkogkkog.support.fixture.domain.CouponFactory.createCoupon;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.JEONG;
 import static com.woowacourse.kkogkkog.support.fixture.domain.MemberFixture.LEO;
 import static com.woowacourse.kkogkkog.support.fixture.domain.WorkspaceFixture.KKOGKKOG;
